scheduler, ci: Add scheduler infrastructure
Add scheduler infrastructure that allows to split
jobs based on cpu intensity. This allows to spread
the CPU load over the whole suite instead of possible
one concentrated place.

The scheduler splits the jobs into two queues with
limit for the CPU intensive being 1/4 of the whole batch.
Is the CPU intensive queue finishes first it will yield
the limit to the regular queue to speed up the rest of the
suite.

Signed-off-by: Ales Musil <amusil@redhat.com>